delete hooks
